+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\T3H\\unit-17\\project\\src\\index.js":"1","D:\\T3H\\unit-17\\project\\src\\App.js":"2","D:\\T3H\\unit-17\\project\\src\\reportWebVitals.js":"3"},{"size":500,"mtime":499162500000,"results":"4","hashOfConfig":"5"},{"size":528,"mtime":499162500000,"results":"6","hashOfConfig":"5"},{"size":362,"mtime":499162500000,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1astmom",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\T3H\\unit-17\\project\\src\\index.js",[],"D:\\T3H\\unit-17\\project\\src\\App.js",[],"D:\\T3H\\unit-17\\project\\src\\reportWebVitals.js",[]]
+[{"D:\\T3H\\unit-17\\project\\src\\index.js":"1","D:\\T3H\\unit-17\\project\\src\\App.js":"2","D:\\T3H\\unit-17\\project\\src\\reportWebVitals.js":"3"},{"size":500,"mtime":499162500000,"results":"4","hashOfConfig":"5"},{"size":1178,"mtime":1610209495820,"results":"6","hashOfConfig":"5"},{"size":362,"mtime":499162500000,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"1astmom",{"filePath":"11","messages":"12","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"D:\\T3H\\unit-17\\project\\src\\index.js",[],["15","16"],"D:\\T3H\\unit-17\\project\\src\\App.js",[],"D:\\T3H\\unit-17\\project\\src\\reportWebVitals.js",[],{"ruleId":"17","replacedBy":"18"},{"ruleId":"19","replacedBy":"20"},"no-native-reassign",["21"],"no-negated-in-lhs",["22"],"no-global-assign","no-unsafe-negation"]
