+++ src/App.css
@@ -1,38 +1,102 @@
-.App {
-  text-align: center;
+@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap");
+
+* {
+    -webkit-box-sizing: border-box;
+    box-sizing: border-box;
+    padding: 0;
+    margin: 0;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+body {
+    font-family: 'Roboto', sans-serif;
+    background-color: #e3e3e3;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
+img {
+    max-width: 100%;
+    max-height: 100%;
 }
 
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+.favorite {
+    background-color: #f8f8f8;
+    max-width: 480px;
+    margin: 15px auto 10px;
 }
 
-.App-link {
-  color: #61dafb;
+.favorite__header {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    padding: 20px 15px;
+    -webkit-box-shadow: 0 1px 15px rgba(137, 137, 137, 0.4);
+    box-shadow: 0 1px 15px rgba(137, 137, 137, 0.4);
+    font-size: 20px;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.favorite__header .fa-arrow-left {
+    width: 80px;
+    margin-right: 15px;
 }
+
+.favorite__header h1 {
+    font-weight: 400;
+    font-size: inherit;
+}
+
+.favorite__header .fa-ellipsis-v {
+    margin-left: auto;
+}
+
+.favorite__playlist__item {
+    padding: 15px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    text-transform: capitalize;
+}
+
+.favorite__playlist__item .thumbnail {
+    width: 80px;
+    height: 80px;
+    border-radius: 6px;
+    overflow: hidden;
+    -webkit-box-shadow: 0 3px 6px 2px rgba(137, 137, 137, 0.4);
+    box-shadow: 0 3px 6px 2px rgba(137, 137, 137, 0.4);
+}
+
+.favorite__playlist__item .thumbnail img {
+    -o-object-fit: cover;
+    object-fit: cover;
+}
+
+.favorite__playlist__item .content {
+    margin-left: 15px;
+    -webkit-box-flex: 1;
+    -ms-flex: 1;
+    flex: 1;
+}
+
+.favorite__playlist__item .content h3 {
+    color: #444444;
+    font-size: 18px;
+    margin-bottom: 5px;
+}
+
+.favorite__playlist__item .content span {
+    font-size: 12px;
+    color: #898989;
+}
+
+.favorite__playlist__item .timeline {
+    margin-right: 10px;
+    font-size: 12px;
+    color: #898989;
+}
+
+/*# sourceMappingURL=App.css.map */
