+++ src/App.scss
@@ -1,0 +1,89 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap');
+
+* {
+    box-sizing: border-box;
+    padding: 0;
+    margin: 0;
+}
+
+body {
+    font-family: 'Roboto', sans-serif;
+    background-color: #e3e3e3;
+}
+
+img {
+    max-width: 100%;
+    max-height: 100%;
+}
+
+.favorite {
+    background-color: #f8f8f8;
+    max-width: 480px;
+    margin: 15px auto 10px;
+
+    &__header {
+        display: flex;
+        align-items: center;
+        padding: 20px 15px;
+        box-shadow: 0 1px 15px rgba(137, 137, 137, 0.4);
+        font-size: 20px;
+
+        .fa-arrow-left {
+            width: 80px;
+            margin-right: 15px;
+        }
+
+        h1 {
+            font-weight: 400;
+            font-size: inherit;
+        }
+
+        .fa-ellipsis-v {
+            margin-left: auto;
+        }
+    }
+
+    &__playlist {
+
+        &__item {
+            padding: 15px;
+            display: flex;
+            align-items: center;
+            text-transform: capitalize;
+
+            .thumbnail {
+                width: 80px;
+                height: 80px;
+                border-radius: 6px;
+                overflow: hidden;
+                box-shadow: 0 3px 6px 2px rgba(137, 137, 137, 0.4);
+
+                img {
+                    object-fit: cover;
+                }
+            }
+
+            .content {
+                margin-left: 15px;
+                flex: 1;
+
+                h3 {
+                    color: #444444;
+                    font-size: 18px;
+                    margin-bottom: 5px;
+                }
+
+                span {
+                    font-size: 12px;
+                    color: #898989;
+                }
+            }
+
+            .timeline {
+                margin-right: 10px;
+                font-size: 12px;
+                color: #898989;
+            }
+        }
+    }
+}
